adapter_requirements.txt
vmware-aria-operations-integration-sdk-lib~=1.1.0
psutil==5.9.4
paramiko==3.5.1
pyvmomi==8.0.3.0.1
